// src/types/TestType.ts
export enum TestType {
  ENTRY = 'ENTRY',
  MODULE_1 = 'MODULE_1',
  MODULE_2 = 'MODULE_2',
  MODULE_3 = 'MODULE_3',
  MODULE_4 = 'MODULE_4',
  MODULE_5 = 'MODULE_5',
  MODULE_6 = 'MODULE_6',
  MODULE_7 = 'MODULE_7',
  MODULE_8 = 'MODULE_8',
  PRACTICE_TASK_1 = 'PRACTICE_TASK_1',
  PRACTICE_TASK_2 = 'PRACTICE_TASK_2'
}